[Flexible] FIXs BaseDeformationmapping & scenes In this PR we:
- remove from the tests 'BezierTetrahedron.scn' as it depend on a plugin.HighOrder

- use msg_info instead of serr in BaseDeformationMapping
The component prints stuff that are really looking like info/debug
messages and I see ne reason why they should be error or warning.
In addition using 'sout/serr' is bad:
- it has an inconsistent naming as it emits warning (while users expect an error).
- it does not add the filename/line from where the message is emitted while msg_info does

So if, in a private fork, you want/need to continue to use the old 'serr' backend...please make your own macro
route msg_info to  the 'old' sout/serr backend instead of the one provided in Sofa.

- Fix examples/material/StVenantKirchoff.scn broken since the merge of In #PR 243 the following change was introduced in MechanicalObject.inl::init():
```cpp
  if (!l_topology && d_useTopology.getValue())
    {
        l_topology.set( this->getContext()->getActiveMeshTopology() );
    }
```

In place of
```cpp
    //Look at a topology associated to this instance of MechanicalObject by a tag
    this->getContext()->get(m_topology, this->getTags());
    // If no topology found, no association, then look to the nearest one
    if(m_topology == NULL)
        m_topology = this->getContext()->getMeshTopology();
```

I assume that the in the scene was the EdgeSetTopologyContainer is not returned by
the getActiveMeshTopology method.

NB: All those implicit linking looks really problematic to me as they are error prone.#PR 243